Bumped to v0.11.5 - Updated with a mix of changes from fd71398c45f183.
Excluding additional promise-based code in index.js that seemed to cause unusual serial errors.
Instead, inserted haltTimedEvents() and resumeTimedEvents() right into loadPropeller().

+++ index.js
@@ -266,15 +266,15 @@ function disconnect() {
     disableWX();
 }
 
-function updateStatus(connected) {
+function updateStatus(connected, socket) {
     if (connected) {
         $('sys-waiting').style.opacity=0.0;
         $('sys-connected').style.opacity=1.0;
-        log('BlocklyProp site connected');
+        log('[S:'+socket.pSocket_.socketId+'] - site connected');
     } else {
         $('sys-waiting').style.opacity=1.0;
         $('sys-connected').style.opacity=0.0;
-        log('BlocklyProp site disconnected');
+        log('[S:'+socket.pSocket_.socketId+'] - site disconnected');
     }
 }
 
@@ -322,13 +322,13 @@ function connect_ws(ws_port, url_path) {
                     serialTerminal(socket, ws_msg.action, ws_msg.portPath, ws_msg.baudrate, ws_msg.msg); // action is "open", "close" or "msg"
                 } else if (ws_msg.type === "port-list-request") {
                     // send an updated port list (and continue on scheduled interval)
-//                  log("Browser requested port-list for socket " + socket.pSocket_.socketId, mDbug);
+                    log('[S:'+socket.pSocket_.socketId+'] -> requested port list', mDbug);
                     sendPortList(socket);
                     portLister.push({socket: socket, scanner: setInterval(function() {sendPortList(socket)}, portListSendInterval)});
                 } else if (ws_msg.type === "hello-browser") {
                     // handle unknown messages
                     helloClient(socket, ws_msg.baudrate || 115200);
-                    updateStatus(true);
+                    updateStatus(true, socket);
                 } else if (ws_msg.type === "debug-cts") {
                     // Handle clear-to-send
                     //TODO Add clear-to-send handling code
@@ -345,7 +345,6 @@ function connect_ws(ws_port, url_path) {
 
         socket.addEventListener('close', function() {
             // Browser socket closed; terminate its port scans and remove it from list of ports.
-            log("Browser socket closing: " + socket.pSocket_.socketId, mDbug);
             let Idx = portLister.findIndex(function(s) {return s.socket === socket});
             if (Idx > -1) {
                 clearInterval(portLister[Idx].scanner);
@@ -353,7 +352,7 @@ function connect_ws(ws_port, url_path) {
             }
             ports.forEach(function(p) {if (p.bSocket === socket) {p.bSocket = null}});
             if (!portLister.length) {
-                updateStatus(false);
+                updateStatus(false, socket);
                 // chrome.app.window.current().drawAttention();  //Disabled to prevent unnecessary user interruption
             }
         });
@@ -393,14 +392,17 @@ function enableWX() {
     }
 }
 
-function disableWX() {
-//Disable wireless port scanning
+function disableWX(retainWXPorts) {
+/* Disable wireless port scanning
+   retainWXPorts [optional] - true = keep list of existing wireless ports; false (default) = delete list of existing wireless ports */
     if(wxScannerInterval) {
         clearInterval(wxScannerInterval);
         wxScannerInterval = null;
     }
-    $('wx-list').innerHTML = '';
-    deleteAllWirelessPorts();
+    if (!retainWXPorts) {
+        $('wx-list').innerHTML = '';
+        deleteAllWirelessPorts();
+    }
 }
 
 function resetWX() {
@@ -409,8 +411,31 @@ function resetWX() {
     wxEnableDelay = setTimeout(enableWX, 500);
 }
 
+function haltTimedEvents() {
+//Halt timed events.  Restart with resumeTimedEvents().
+    return new Promise(function(resolve) {
+        //Disable wired and wireless port scanning
+        log('Halting timed events', mDbug);
+        disableW();
+        disableWX(true);
+        resolve();
+    })
+}
+
+function resumeTimedEvents() {
+//Resume timed events that were stopped via haltTimedEvents().
+    return new Promise(function(resolve) {
+        //Enable wired and wireless port scanning
+        log('Resuming timed events', mDbug);
+        enableW();
+        enableWX();
+        resolve();
+    })
+}
+
 function scanWPorts() {
 // Generate list of current wired ports (filtered according to platform and type)
+    log('Scanning wired ports', mDbug);
     chrome.serial.getDevices(
         function(portlist) {
             let wn = [];
@@ -431,6 +456,7 @@ function scanWPorts() {
 
 function scanWXPorts() {
 // Generate list of current wireless ports
+    log('Scanning wireless ports', mDbug);
     discoverWirelessPorts();
     ageWirelessPorts();
 }
@@ -439,7 +465,7 @@ function sendPortList(socket) {
 // Find and send list of communication ports (filtered according to platform and type) to browser via socket
     let wn = [];
     let wln = [];
-//    log("sendPortList() for socket " + socket.pSocket_.socketId, mDbug);
+    log('[S:'+socket.pSocket_.socketId+'] <- Sending port list', mDbug);
     // gather separated and sorted port lists (wired names and wireless names)
     ports.forEach(function(p) {if (p.isWired) {wn.push(p.name)} else {wln.push(p.name)}});
     wn.sort();
